+++ backend/core/src/main/java/site/timecapsulearchive/core/domain/capsule/repository/GroupCapsuleOpenRepository.java
@@ -1,0 +1,8 @@
+package site.timecapsulearchive.core.domain.capsule.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import site.timecapsulearchive.core.domain.capsule.entity.GroupCapsuleOpen;
+
+public interface GroupCapsuleOpenRepository extends JpaRepository<GroupCapsuleOpen, Long> {
+
+}
