--- backend/core/src/main/java/site/timecapsulearchive/core/domain/groupcapsule/repository/GroupCapsuleOpenRepository.java
@@ -1,8 +1,0 @@
-package site.timecapsulearchive.core.domain.groupcapsule.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import site.timecapsulearchive.core.domain.groupcapsule.entity.GroupCapsuleOpen;
-
-public interface GroupCapsuleOpenRepository extends JpaRepository<GroupCapsuleOpen, Long> {
-
-}
